Homework1/Client/requirements.txt
grpcio 
grpcio-tools
yfinance
mysql-connector-python
email-validator